requirements.txt
browser-use>=0.12.0
langchain-google-genai>=4.0.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0
Pillow>=10.0

# Portal (agentchattr fork)
fastapi>=0.110
uvicorn[standard]>=0.29
mcp>=1.0
